+++ plugin.yml
@@ -2,7 +2,7 @@ name : KickAll
 author: AmmyR
 api: 3.0.0
 main: AmmyR\KickAll\Main
-version: 1.2.0
+version: 1.2.2
 permissions:
   kickall.command:
     description: "Allows to use /kickall command."
