+++ src/AmmyR/KickAll/Main.php
@@ -8,12 +8,11 @@ Twitter: @AmmyRQ
 namespace AmmyR\KickAll;
 
 use pocketmine\plugin\PluginBase;
-use AmmyR\KickAll\Command;
 
 class Main extends PluginBase {
 
 	public function onEnable() : void{
-		$this->getServer()->getCommandMap()->register("kickall", new Command($this));
+		$this->getServer()->getCommandMap()->register("kickall", new KickCommand($this));
 	}
 
 }
